backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pandas==2.2.0
python-multipart==0.0.6
python-dotenv==1.0.1
langchain==0.1.16
langchain-community==0.1.0
langchain-google-genai==1.0.5
pydantic==2.5.0
pydantic-settings==2.1.0
